remove and rename unused columns, add password_encryption_type
