+++ database/entity/0052-change_users_password_encryption/User.ts
@@ -115,7 +115,6 @@ export class User extends BaseEntity {
   @Column({ name: 'publisher_id', default: 0 })
   publisherId: number
 
-
   @OneToMany(() => Contribution, (contribution) => contribution.user)
   @JoinColumn({ name: 'user_id' })
   contributions?: Contribution[]
