+++ database/entity/User.ts
@@ -1,1 +1,1 @@
-export { User } from './0052-change_user_password_encryption/User'
+export { User } from './0052-change_users_password_encryption/User'
